requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
scikit-learn>=1.2.2
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.8.0
numpy>=1.24.3
pytest>=7.3.1
httpx>=0.24.0
python-multipart>=0.0.6
joblib>=1.2.0
datasets==3.4.0
pandas==2.2.3